main/requirements.txt
scipy
tenacity
tiktoken
termcolor
openai
langchain
ipylangchat
langchain_openai
langchain_community
langchain_huggingface
langchain_chroma
langchain_text_splitters
urllib3
pymupdf
anywidget
uvicorn
starlette
ome-zarr
ome_types
aicsimageio
vizarr
anndata
zarr
esbuild_py 
deepmerge
python-dotenv
duckdb
jupyterlab
vitessce==3.3.4